requirements.txt
pytorch-cuda>=11.8
pytorch>=2.0.0
torchvision
numpy
lightning
ruff
loguru
wandb
h5py
Pillow
pydantic
pydantic-settings